+++ Content/Items/CapturedMonster.cs
@@ -5,6 +5,7 @@ using Terraria.ModLoader.IO;
 using Microsoft.Xna.Framework;
 using Terraria.Audio;
 using Monstermon.Content.Buffs;
+using Monstermon.Content.Systems;
 
 namespace Monstermon.Content.Items
 {
